requirements.txt
alpaca-py
supabase
pandas
python-dotenv
requests
lxml
